Integrated real hasher/signer/verifier, into transport/commitment test.  It works!
git-svn-id: https://svn.mpi-sws.org/DS/freepastry/trunk@4490 c21dacaf-8602-0410-aacd-8bc41df2cbe7

+++ pastry/src/org/mpisws/p2p/transport/peerreview/history/logentry/EvtRecv.java
@@ -59,7 +59,7 @@ public class EvtRecv<Handle extends RawSerializable> extends HistoryEvent {
   }
   
   public EvtRecv(Handle senderHandle, long topSeq, ByteBuffer payload,
-      short relevantLen, HashProvider hasher) {
+      int relevantLen, HashProvider hasher) {
     this.senderHandle = senderHandle;
     this.senderSeq = topSeq;
     this.payload = new byte[payload.remaining()];
